+++ personal_mnemonic_medium/destination/anki_prompt_base.py
