+++ personal_mnemonic_medium/destination/destination_commands.py
@@ -1,0 +1,17 @@
+from collections.abc import Sequence
+from dataclasses import dataclass
+
+from ..source.prompt_base import BasePrompt, DestinationPrompt
+
+
+@dataclass(frozen=True)
+class DeletePrompts:
+    prompts: Sequence[DestinationPrompt]
+
+
+@dataclass(frozen=True)
+class PushPrompts:
+    prompts: Sequence[BasePrompt]
+
+
+PromptDestinationCommand = DeletePrompts | PushPrompts
