--- personal_mnemonic_medium/domain/prompt_destination/destination_commands.py
@@ -1,17 +1,0 @@
-from collections.abc import Sequence
-from dataclasses import dataclass
-
-from ..prompts.base_prompt import BasePrompt, DestinationPrompt
-
-
-@dataclass(frozen=True)
-class DeletePrompts:
-    prompts: Sequence[DestinationPrompt]
-
-
-@dataclass(frozen=True)
-class PushPrompts:
-    prompts: Sequence[BasePrompt]
-
-
-PromptDestinationCommand = DeletePrompts | PushPrompts
