--- personal_mnemonic_medium/domain/prompt_source/prompt_extractors/base_prompt_extractor.py
@@ -1,13 +1,0 @@
-from collections.abc import Sequence
-from typing import Protocol
-
-from ...prompts.base_prompt import BasePrompt
-from ..document_ingesters.document import Document
-
-
-# TODO: https://github.com/MartinBernstorff/personal-mnemonic-medium/issues/297 Implement ClozeExtractor
-class BasePromptExtractor(Protocol):
-    def extract_prompts(
-        self, document: Document
-    ) -> Sequence[BasePrompt]:
-        ...
