--- personal_mnemonic_medium/domain/prompt_source/prompt_extractors/test_cloze_prompt_extractor.py
@@ -1,29 +1,0 @@
-from pathlib import Path
-
-from ..document_ingesters.document import Document
-from .cloze_prompt_extractor import ClozePromptExtractor
-
-
-def test_cloze_prompt_extractor(tmpdir: Path):
-    doc = Document(
-        content=r"""
-
-What is the meaning of life? {42}
-
-
-This is another block without any cloze prompts.
-
-#anki/tag/test_tag
-
-""",
-        source_path=tmpdir / "test.md",
-    )
-
-    extractor = ClozePromptExtractor().extract_prompts(doc)
-
-    assert len(extractor) == 1
-    assert (
-        extractor[0].text
-        == r"What is the meaning of life? {{c734::42}}"
-    )
-    assert extractor[0].tags == ["anki/tag/test_tag"]
