+++ personal_mnemonic_medium/source/document.py
