+++ personal_mnemonic_medium/source/ingester_base.py
