+++ personal_mnemonic_medium/source/prompt_base.py
