+++ personal_mnemonic_medium/utils/test_hash_cleaned_str.py
